Move get_or_compute_tag_hash() to MetadataManager
